구현 진도표에 맞춰 Class명 변경 및 새로운 클래스 생성
