+++ src/MemberManage/UserController.java
@@ -7,12 +7,12 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.time.LocalDate;
 
-public class MemberManagement {
+public class UserController {
     private Member member = null;
     private ConnectMyDB connectMyDB = null;
     private ResultSet rs = null;
 
-    public MemberManagement() throws SQLException, ClassNotFoundException {
+    public UserController() throws SQLException, ClassNotFoundException {
         member = new Member();
         connectMyDB = new ConnectMyDB();
     }
@@ -61,28 +61,4 @@ public class MemberManagement {
         return rs.next();
     }
 
-    public Member getMemberInfo(String id) throws SQLException {
-        String query = "select * from member where memberID = ?";
-        PreparedStatement pstmt = connectMyDB.getConnection().prepareStatement(query);
-        pstmt.setString(1, id);
-
-        rs = pstmt.executeQuery();
-        if(rs.next()) {
-            String name = rs.getString("name");
-            int age = rs.getInt("age");
-            String date = rs.getString("birthday");
-            String phoneNumber = rs.getString("phoneNumber");
-
-            connectMyDB.disConnectMyDB();
-            return new Member(id, name, age, date, phoneNumber);
-        }
-        try{
-            if(connectMyDB.getConnection() != null) connectMyDB.getConnection().close();
-            if(pstmt != null) pstmt.close();
-            if(rs != null) rs.close();
-        } catch(Exception e){
-            e.printStackTrace();
-        }
-        return null;
-    }
 }
